moving from poultry/enter-site-name to poultry/site-name
