+++ app/constants/routes.js
@@ -172,7 +172,7 @@ export const poultryClaimRoutes = {
   confirmation: "/poultry/confirmation",
   costOfChanges: "/poultry/cost-of-changes",
   dateOfVisit: "/poultry/date-of-visit",
-  enterSiteName: "/poultry/enter-site-name",
+  enterSiteName: "/poultry/site-name",
   enterCphNumber: "/poultry/enter-cph-number",
   enterSiteDetails: "/poultry/enter-site-details",
   interview: "/poultry/interview",
@@ -195,7 +195,7 @@ export const poultryClaimViews = {
   costOfChanges: "poultry/claim/biosecurity-cost-of-changes",
   dateOfVisit: "poultry/claim/date-of-visit",
   interview: "poultry/claim/interview",
-  enterSiteName: "poultry/claim/enter-site-name",
+  enterSiteName: "poultry/claim/site-name",
   enterCphNumber: "poultry/claim/enter-cph-number",
   minimumNumberOfBirds: "poultry/claim/minimum-number-of-birds",
   minimumNumberOfBirdsException: "poultry/claim/minimum-number-of-birds-exception",
