+++ test/integration/narrow/routes/poultry/claim/select-the-site.test.js
@@ -546,7 +546,7 @@ describe("/poultry/select-site", () => {
       crumb = await getCrumbs(server);
     });
 
-    test("redirects to enter-site-name when selecting new site from single site view", async () => {
+    test("redirects to site-name when selecting new site from single site view", async () => {
       when(getSessionData)
         .calledWith(expect.anything(), sessionEntryKeys.poultryClaim)
         .mockReturnValue({
@@ -577,7 +577,7 @@ describe("/poultry/select-site", () => {
       });
 
       expect(res.statusCode).toBe(302);
-      expect(res.headers.location).toEqual("/poultry/enter-site-name");
+      expect(res.headers.location).toEqual("/poultry/site-name");
 
       expect(setSessionData).toHaveBeenCalledWith(
         expect.anything(),
@@ -616,7 +616,7 @@ describe("/poultry/select-site", () => {
       );
     });
 
-    test("redirects to enter-site-name when selecting new site from multiple sites view", async () => {
+    test("redirects to site-name when selecting new site from multiple sites view", async () => {
       when(getSessionData)
         .calledWith(expect.anything(), sessionEntryKeys.poultryClaim)
         .mockReturnValue({
@@ -659,7 +659,7 @@ describe("/poultry/select-site", () => {
       });
 
       expect(res.statusCode).toBe(302);
-      expect(res.headers.location).toEqual("/poultry/enter-site-name");
+      expect(res.headers.location).toEqual("/poultry/site-name");
       expect(setSessionData).toHaveBeenCalledWith(
         expect.anything(),
         sessionEntryKeys.poultryClaim,
